+++ src/Ev.ServiceBus.Abstractions/Configuration/MessageReceptionBuilder.cs
@@ -6,6 +6,10 @@ public class MessageReceptionBuilder
 {
     private readonly MessageReceptionRegistration _registration;
 
+    public Type HandlerType => _registration.HandlerType;
+
+    public Type PayloadType => _registration.PayloadType;
+
     public MessageReceptionBuilder(ClientOptions clientOptions, Type payloadType, Type handlerType)
     {
         _registration = new MessageReceptionRegistration(clientOptions, payloadType, handlerType);