+++ src/Ev.ServiceBus/Reception/MessageReceptionHandler.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Ev.ServiceBus.Abstractions;
 using Ev.ServiceBus.Abstractions.MessageReception;
+using Ev.ServiceBus.Exceptions;
 using Ev.ServiceBus.Management;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -16,7 +17,7 @@ public class MessageReceptionHandler
 {
     private readonly MethodInfo _callHandlerInfo;
     private readonly IMessagePayloadSerializer _messagePayloadSerializer;
-    private readonly ILogger<MessageReceptionHandler> _logger;
+    private readonly ILogger<LoggingExtensions.MessageProcessing> _logger;
     private readonly MessageMetadataAccessor _messageMetadataAccessor;
     private readonly IEnumerable<IServiceBusEventListener> _eventListeners;
     private readonly IServiceProvider _provider;
@@ -24,7 +25,7 @@ public class MessageReceptionHandler
     public MessageReceptionHandler(
         IServiceProvider provider,
         IMessagePayloadSerializer messagePayloadSerializer,
-        ILogger<MessageReceptionHandler> logger,
+        ILogger<LoggingExtensions.MessageProcessing> logger,
         IMessageMetadataAccessor messageMetadataAccessor,
         IEnumerable<IServiceBusEventListener> eventListeners)
     {
@@ -38,16 +39,7 @@ public class MessageReceptionHandler
 
     public async Task HandleMessageAsync(MessageContext context)
     {
-        var scopeValues = new Dictionary<string, string>
-        {
-            ["EVSB_Client"] = context.ClientType.ToString(),
-            ["EVSB_ResourceId"] = context.ResourceId,
-            ["EVSB_MessageId"] = context.Message.MessageId,
-            ["EVSB_SessionId"] = context.SessionArgs?.SessionId ?? "none",
-            ["EVSB_PayloadTypeId"] = context.PayloadTypeId ?? "none",
-            ["EVSB_ReceptionHandler"] = context.ReceptionRegistration?.HandlerType.FullName ?? "none"
-        };
-        using (_logger.BeginScope(scopeValues))
+        using (AddLoggingContext(context))
         {
             _messageMetadataAccessor.SetData(context);
 
@@ -56,54 +48,42 @@ public class MessageReceptionHandler
             {
                 await listener.OnExecutionStart(executionStartedArgs);
             }
-            _logger.LogInformation("[Ev.ServiceBus] New message received from {EVSB_Client} '{EVSB_ResourceId}' : {EVSB_MessageLabel}",
-                context.ClientType,
-                context.ResourceId,
-                context.Message.Subject);
 
             var sw = new Stopwatch();
             sw.Start();
             try
             {
                 if (context.PayloadTypeId == null)
-                {
                     throw new MessageIsMissingPayloadTypeIdException(context);
-                }
 
                 if (context.ReceptionRegistration == null)
-                {
                     return;
-                }
 
                 if (context.CancellationToken.IsCancellationRequested)
-                {
-                    _logger.LogInformation("[Ev.ServiceBus] Stopping the execution because cancellation was requested");
                     return;
-                }
-
-                _logger.LogDebug(
-                    "[Ev.ServiceBus] Executing {EVSB_PayloadTypeId}:{EVSB_ReceptionHandler} handler",
-                    context.ReceptionRegistration.PayloadTypeId,
-                    context.ReceptionRegistration.HandlerType.FullName);
 
                 var @event = _messagePayloadSerializer.DeSerializeBody(context.Message.Body.ToArray(), context.ReceptionRegistration!.PayloadType);
                 var methodInfo = _callHandlerInfo.MakeGenericMethod(context.ReceptionRegistration.PayloadType);
                 await ((Task) methodInfo.Invoke(this, new[] { context.ReceptionRegistration, @event, context.CancellationToken })!);
-
-                _logger.LogInformation(
-                    "[Ev.ServiceBus] Execution of {EVSB_PayloadTypeId}:{EVSB_ReceptionHandler} reception handler successful in {EVSB_Duration} milliseconds",
-                    context.ReceptionRegistration.PayloadTypeId,
-                    context.ReceptionRegistration.HandlerType.FullName,
-                    sw.ElapsedMilliseconds);
             }
-            catch (Exception ex) when (LogError(ex))
+            catch (Exception ex)
             {
                 var executionFailedArgs = new ExecutionFailedArgs(context, ex);
                 foreach (var listener in _eventListeners)
                 {
                     await listener.OnExecutionFailed(executionFailedArgs);
                 }
-                throw;
+
+                var executionContext = context.ReadExecutionContext();
+
+                throw new FailedToProcessMessageException(
+                    clientType: executionContext.ClientType,
+                    resourceId: executionContext.ResourceId,
+                    messageId: executionContext.MessageId,
+                    payloadTypeId: executionContext.PayloadTypeId,
+                    sessionId: executionContext.SessionId,
+                    handlerName: executionContext.HandlerName,
+                    ex);
             }
             finally
             {
@@ -115,8 +95,22 @@ public class MessageReceptionHandler
             {
                 await listener.OnExecutionSuccess(executionSucceededArgs);
             }
-            _logger.LogInformation("[Ev.ServiceBus] Message finished execution in {EVSB_Duration} milliseconds", sw.ElapsedMilliseconds);
+            _logger.MessageExecutionCompleted(sw.ElapsedMilliseconds);
         }
+    }
+
+    private IDisposable? AddLoggingContext(MessageContext context)
+    {
+        var executionContext = context.ReadExecutionContext();
+
+        return _logger.ProcessingInProgress(
+            clientType: executionContext.ClientType,
+            resourceId: executionContext.ResourceId,
+            messageId: executionContext.MessageId,
+            payloadTypeId: executionContext.PayloadTypeId,
+            sessionId: executionContext.SessionId,
+            handlerName: executionContext.HandlerName
+        );
     }
 
     private async Task CallHandler<TMessagePayload>(
@@ -127,17 +121,5 @@ public class MessageReceptionHandler
         var handler = (IMessageReceptionHandler<TMessagePayload>) _provider.GetRequiredService(messageReceptionRegistration.HandlerType);
 
         await handler.Handle(@event, token);
-    }
-
-    /// <summary>
-    ///     workaround to attach the log scope to the logged exception
-    ///     https://andrewlock.net/how-to-include-scopes-when-logging-exceptions-in-asp-net-core/
-    /// </summary>
-    /// <param name="ex"></param>
-    /// <returns></returns>
-    private bool LogError(Exception ex)
-    {
-        _logger.LogError(ex, ex.Message);
-        return true;
     }
 }